Create useAvatar for getting avatar data
