+++ src/hooks/useAvatar.ts
@@ -1,24 +1,10 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { supabase } from '../utils/supabaseClient';
-import Image from 'next/image';
 
-export default function Avatar({
-  url,
-  size,
-  onUpload,
-}: {
-  url: string | null;
-  size: number;
-  onUpload: (url: string) => void;
-}) {
+export default function useAvatar({ onUpload }: { onUpload: (url: string) => void }) {
   const [avatarUrl, setAvatarUrl] = useState<string | null>(null);
-  const [uploading, setUploading] = useState(false);
-
-  useEffect(() => {
-    if (url) downloadImage(url);
-  }, [url]);
-
-  async function downloadImage(path: string) {
+  const [avatarUploading, setUploading] = useState(false);
+  async function downloadAvatar(path: string) {
     try {
       const { data, error } = await supabase.storage.from('avatars').download(path);
       if (error) {
@@ -52,37 +38,18 @@ export default function Avatar({
       if (uploadError) {
         throw uploadError;
       }
-
       onUpload(filePath);
+      downloadAvatar(filePath);
     } catch (error) {
       if (error instanceof Error) alert(error.message);
     } finally {
       setUploading(false);
     }
   }
-  return (
-    <div>
-      {avatarUrl ? (
-        <Image src={avatarUrl} alt='Avatar' className='avatar image' width={size} height={size} />
-      ) : (
-        <Image src='/images/cat.png' className='avatar no-image' height={size} width={size} />
-      )}
-      <div style={{ width: size }}>
-        <label className='button primary block' htmlFor='single'>
-          {uploading ? 'Uploading ...' : 'Upload'}
-        </label>
-        <input
-          style={{
-            visibility: 'hidden',
-            position: 'absolute',
-          }}
-          type='file'
-          id='single'
-          accept='image/*'
-          onChange={uploadAvatar}
-          disabled={uploading}
-        />
-      </div>
-    </div>
-  );
+  return {
+    avatarUploading,
+    avatarUrl,
+    downloadAvatar,
+    uploadAvatar,
+  };
 }
